pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.0",
    "cryptography>=46.0.2",
    "fastapi[standard]>=0.118.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.27",
    "structlog>=25.4.0",
]
